+++ serverCommunicator/src/main/java/com/doodeec/utils/network/listener/GSONRequestListener.java
@@ -1,7 +1,0 @@
-package com.doodeec.utils.network.listener;
-
-/**
- * @author Dusan Bartos
- */
-public interface GSONRequestListener<GSONType> extends BaseRequestListener<GSONType>{
-}